server/requirements.txt
fastapi == 0.70.0
uvicorn == 0.16.0
motor==2.5.1
xmltodict==0.12.0
python3-saml == 1.14.0
python-multipart == 0.0.5
itsdangerous == 2.1.2
starlette-authlib == 0.1.17
